nn structure, module improve, and refactor
