+++ src/lib.rs
@@ -1,9 +1,11 @@
 pub mod backward;
 pub mod data;
 pub mod linalg;
+pub mod module;
 pub mod nn;
 mod op;
 pub mod tensor;
 mod tensor_data;
 
+// define short paths
 pub use tensor::Tensor;
